neetcode_150/stack/src/min_stack.py
#STATS
#   Runtime         Memory
#   O(1)            O(n)
# Min Stack #155

class MinStack:
    def __init__(self):
        self.stack = []
        self.min_stack = []

    def push(self, val: int) -> None:
        self.stack.append(val)
        if self.min_stack and val <= self.min_stack[-1]:
            self.min_stack.append(val)
        elif not self.min_stack:
            self.min_stack.append(val)

    def pop(self) -> None:
        if self.stack[-1] == self.min_stack[-1]:
            self.min_stack.pop()
        self.stack.pop()


    def top(self) -> int:
        return self.stack[-1]

    def getMin(self) -> int:
        return self.min_stack[-1]
